Fixes day2 worker playbook for disconnected installs
Currently the day2 worker playbook is broken for disconnected installs
as the patching logic for mirrored registries does not exist. This
commit fixes this although its mostly a copy of the templates from
the main create cluster role so probably can be done a lot nicer.
Also, the CSR approval mechanism is still broken.