Taking into account fully qualified image URLs
You can specify the image as `$REGISTRY/$NS/$IMAGE:$TAG` but you can
also specify an image as `$REGISTRY/$NS/$IMAGE@$SHA_ALGO:$SHA_STR` e.g.
how they are set in `release.txt`: "Pull From: $IMAGE"

If the image is fully qualified then there will be an `@` character in
the URL, if that happens then there is no need to re-inspect the image
to get the signature.